+++ src/main/java/com/uberkautilya/splitwise_lld3/dtos/RequestDto.java
@@ -6,5 +6,5 @@ import lombok.Setter;
 @Getter
 @Setter
 public class RequestDto {
-    int groupId;
+    Long groupId;
 }
